+++ src/extension.ts
@@ -18,7 +18,7 @@ import { testApi } from "./commands/test-api";
 import { initializeLocalStorage } from "./utils/localStorage";
 import { useRecipe } from "./commands/use-recipe";
 import { createRecipe } from "./commands/create-recipe";
-import { providesCodeCompletion } from "./code-completion/assistant-completion";
+import { providesCodeCompletion } from "./features/assistant-completion";
 import { useRecipeCallback } from "./graphql-api/use-recipe";
 import { UriHandler } from "./utils/uriHandler";
 import { getUser } from "./graphql-api/user";
@@ -35,7 +35,7 @@ import {
   showCodigaWebview,
   updateWebview,
 } from "./commands/webview";
-import { provideInlineComplextion } from "./code-completion/inline-completion";
+import { provideInlineComplextion } from "./features/inline-completion";
 import { AssistantRecipe } from "./graphql-api/types";
 import { subscribeToDocumentChanges } from "./diagnostics/diagnostics";
 import { applyFix, RosieFixAction } from "./rosie/rosiefix";
