+++ src/features/assistant-completion.ts
