+++ src/features/inline-completion.ts
